backend/app/api/routes/esclavo/requirements.txt
fastapi
pydantic
asyncpg
python-dotenv
uvicorn